// libs-anilist.old/anilist/src/generated/types.ts
export interface fetchUserLists_MediaListCollection_user {
  __typename: "User";
  /**
   * The id of the user
   */
  id: number;
}

export interface fetchUserLists_MediaListCollection_lists_entries_startedAt {
  __typename: "FuzzyDate";
  /**
   * Numeric Year (2017)
   */
  year: number | null;
  /**
   * Numeric Month (3)
   */
  month: number | null;
  /**
   * Numeric Day (24)
   */
  day: number | null;
}

export interface fetchUserLists_MediaListCollection_lists_entries_completedAt {
  __typename: "FuzzyDate";
  /**
   * Numeric Year (2017)
   */
  year: number | null;
  /**
   * Numeric Month (3)
   */
  month: number | null;
  /**
   * Numeric Day (24)
   */
  day: number | null;
}
// ...
export interface fetchUserLists_MediaListCollection_lists_entries_media {
  __typename: "Media";
  /**
   * The id of the media
   */
  id: number;
  /**
   * The current releasing status of the media
   */
  status: MediaStatus | null;
  /**
   * The format the media was released in
   */
  format: MediaFormat | null;
  /**
   * The amount of episodes the anime has when complete
   */
  episodes: number | null;
  /**
   * The general length of each anime episode in minutes
   */
  duration: number | null;
  /**
   * Short description of the media's story and characters
   */
  description: string | null;
  /**
   * The cover images of the media
   */
  coverImage: fetchUserLists_MediaListCollection_lists_entries_media_coverImage | null;
  /**
   * The first official release date of the media
   */
  startDate: fetchUserLists_MediaListCollection_lists_entries_media_startDate | null;
  /**
   * The last official release date of the media
   */
  endDate: fetchUserLists_MediaListCollection_lists_entries_media_endDate | null;
  /**
   * The official titles of the media in various languages
   */
  title: fetchUserLists_MediaListCollection_lists_entries_media_title | null;
  /**
   * Other media in the same or connecting franchise
   */
  relations: fetchUserLists_MediaListCollection_lists_entries_media_relations | null;
}

export interface fetchUserLists_MediaListCollection_lists_entries {
  __typename: "MediaList";
  /**
   * The id of the list entry
   */
  id: number;
  /**
   * The watching/reading status
   */
  status: MediaListStatus | null;
  /**
   * Text notes
   */
  notes: string | null;
  /**
   * The amount of episodes/chapters consumed by the user
   */
  progress: number | null;
  /**
   * When the entry was started by the user
   */
  startedAt: fetchUserLists_MediaListCollection_lists_entries_startedAt | null;
  /**
   * When the entry was completed by the user
   */
  completedAt: fetchUserLists_MediaListCollection_lists_entries_completedAt | null;
  media: fetchUserLists_MediaListCollection_lists_entries_media | null;
}

export interface fetchUserLists_MediaListCollection_lists {
  __typename: "MediaListGroup";
  name: string | null;
  /**
   * Media list entries
   */
  entries: (fetchUserLists_MediaListCollection_lists_entries | null)[] | null;
}

export interface fetchUserLists_MediaListCollection {
  __typename: "MediaListCollection";
  /**
   * The owner of the list
   */
  user: fetchUserLists_MediaListCollection_user | null;
  /**
   * Grouped media list entries
   */
  lists: (fetchUserLists_MediaListCollection_lists | null)[] | null;
}

export interface fetchUserLists {
  /**
   * Media list collection query, provides list pre-grouped by status & custom lists. User ID and Media Type arguments required.
   */
  MediaListCollection: fetchUserLists_MediaListCollection | null;
}

export interface fetchUserListsVariables {
  userId: number;
  type: MediaType;
  statuses: MediaListStatus[];
}

/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: findUsersByName
// ====================================================

export interface findUsersByName_Page_users_avatar {
  __typename: "UserAvatar";
  /**
   * The avatar of user at its largest size
   */
  large: string | null;
  /**
   * The avatar of user at medium size
   */
  medium: string | null;
}

export interface findUsersByName_Page_users {
  __typename: "User";
  /**
   * The id of the user
   */
  id: number;
  /**
   * The name of the user
   */
  name: string;
  /**
   * The user's avatar images
   */
  avatar: findUsersByName_Page_users_avatar | null;
  /**
   * The user's banner images
   */
  bannerImage: string | null;
}

export interface findUsersByName_Page {
  __typename: "Page";
  users: (findUsersByName_Page_users | null)[] | null;
}

export interface findUsersByName {
  Page: findUsersByName_Page | null;
}

export interface findUsersByNameVariables {
  query: string;
  page: number;
  perPage: number;
}
// ...
export interface getUserById_User_avatar {
  __typename: "UserAvatar";
  /**
   * The avatar of user at its largest size
   */
  large: string | null;
  /**
   * The avatar of user at medium size
   */
  medium: string | null;
}

export interface getUserById_User {
  __typename: "User";
  /**
   * The id of the user
   */
  id: number;
  /**
   * The name of the user
   */
  name: string;
  /**
   * The user's avatar images
   */
  avatar: getUserById_User_avatar | null;
  /**
   * The user's banner images
   */
  bannerImage: string | null;
}

export interface getUserById {
  /**
   * User query
   */
  User: getUserById_User | null;
}

export interface getUserByIdVariables {
  userId: number;
}

/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: getUserLists
// ====================================================

export interface getUserLists_MediaListCollection_user {
  __typename: "User";
  /**
   * The id of the user
   */
  id: number;
}
// ...
export interface getUserLists_MediaListCollection_lists {
  __typename: "MediaListGroup";
  name: string | null;
  /**
   * Media list entries
   */
  entries: (getUserLists_MediaListCollection_lists_entries | null)[] | null;
}

export interface getUserLists_MediaListCollection {
  __typename: "MediaListCollection";
  /**
   * The owner of the list
   */
  user: getUserLists_MediaListCollection_user | null;
  /**
   * Grouped media list entries
   */
  lists: (getUserLists_MediaListCollection_lists | null)[] | null;
}

export interface getUserLists {
  /**
   * Media list collection query, provides list pre-grouped by status & custom lists. User ID and Media Type arguments required.
   */
  MediaListCollection: getUserLists_MediaListCollection | null;
}

export interface getUserListsVariables {
  userId: number;
  type: MediaType;
  statuses: MediaListStatus[];
}
// ...
export enum MediaListStatus {
  COMPLETED = "COMPLETED",
  CURRENT = "CURRENT",
  DROPPED = "DROPPED",
  PAUSED = "PAUSED",
  PLANNING = "PLANNING",
  REPEATING = "REPEATING",
}
// ...
export enum MediaType {
  ANIME = "ANIME",
  MANGA = "MANGA",
}